agents/requirements.txt
langchain-core>=0.3
langchain-openai>=0.3
pydantic>=2.0
python-dotenv
watchdog>=3.0

# Pixelsmith (art generation)
torch>=2.0
diffusers>=0.25
transformers>=4.36,<5.0
accelerate
rembg
Pillow>=10.0
opencv-python-headless
safetensors
numpy
